Seperate decompiler statements
