+++ UndertaleModLib/Decompiler/Decompiler.cs
@@ -1,8 +1,6 @@
 using System;
-using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Collections.Specialized;
 using System.Diagnostics;
 using System.Globalization;
 using System.IO;
@@ -11,183 +9,10 @@ using System.Text;
 using System.Threading.Tasks;
 using UndertaleModLib.Models;
 using UndertaleModLib.Util;
-using static UndertaleModLib.Decompiler.Decompiler;
 
 namespace UndertaleModLib.Decompiler
 {
-    /// <summary>
-    /// The DecompileContext is global for the entire decompilation run, or possibly multiple runs. It caches the decompilation results which don't change often
-    /// to speedup decompilation.
-    /// </summary>
-    public class GlobalDecompileContext
-    {
-        public UndertaleData Data;
-
-        public bool EnableStringLabels;
-
-        public List<string> DecompilerWarnings = new List<string>();
-
-        /// <summary>
-        /// A cache of resolved function argument types. This is kept here because decompiling is slow, and there is no need to do it every time
-        /// unless the code has changed.
-        /// </summary>
-        public Dictionary<string, AssetIDType[]> ScriptArgsCache = new Dictionary<string, AssetIDType[]>();
-
-        /// <summary>
-        /// A cache of function to actual name mapping. GMS2.3+ sometimes (usually when dealing with global scripts) calls method functions
-        /// using the legacy call operator, passing the anonymous function directly. This dictionary contains a map from UndertaleFunction
-        /// to its actual name, obtained by decompiling the parent CodeObject and looking for the assignment to global variable with function
-        /// name.
-        /// </summary>
-        public Dictionary<UndertaleFunction, string> AnonymousFunctionNameCache = new Dictionary<UndertaleFunction, string>();
-
-        public GlobalDecompileContext(UndertaleData data, bool enableStringLabels)
-        {
-            this.Data = data;
-            this.EnableStringLabels = enableStringLabels;
-        }
-
-        public void ClearDecompilationCache()
-        {
-            // This will not be done automatically, because it would cause significant slowdown having to recalculate this each time, and there's no reason to reset it if it's decompiling a bunch at once.
-            // But, since it is possible to invalidate this data, we add this here so we'll be able to invalidate it if we need to.
-            ScriptArgsCache.Clear();
-            AnonymousFunctionNameCache.Clear();
-        }
-    }
-
-    /// <summary>
-    /// The DecompileContext is bound to the currently decompiled code block
-    /// </summary>
-    public class DecompileContext
-    {
-        public GlobalDecompileContext GlobalContext;
-        public UndertaleCode TargetCode;
-        public UndertaleGameObject Object;
-        public static bool GMS2_3;
-        public bool AssetResolutionEnabled => !GlobalContext.Data.IsVersionAtLeast(2023, 8);
-
-        public DecompileContext(GlobalDecompileContext globalContext, UndertaleCode code, bool computeObject = true)
-        {
-            GlobalContext = globalContext;
-            TargetCode = code;
-
-            if (code.ParentEntry != null)
-                throw new InvalidOperationException("This code block represents a function nested inside " + code.ParentEntry.Name + " - decompile that instead");
-
-            if (computeObject && globalContext.Data is not null)
-            {
-                // TODO: This is expensive, move it somewhere else as a dictionary
-                // and have it update when events/objects are modified.
-                
-                // Currently using for loops on purpose, as foreach has memory issues due to IEnumerable
-                for (int i = 0; i < globalContext.Data.GameObjects.Count; i++)
-                {
-                    UndertaleGameObject obj = globalContext.Data.GameObjects[i];
-                    for (int j = 0; j < obj.Events.Count; j++)
-                    {
-                        var eventList = obj.Events[j];
-                        for (int k = 0; k < eventList.Count; k++)
-                        {
-                            UndertaleGameObject.Event subEvent = eventList[k];
-                            for (int l = 0; l < subEvent.Actions.Count; l++)
-                            {
-                                UndertaleGameObject.EventAction ev = subEvent.Actions[l];
-                                if (ev.CodeId != code) continue;
-                                Object = obj;
-                                return;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        #region Struct management
-        public List<Expression> ArgumentReplacements;
-        public bool DecompilingStruct;
-        #endregion
-
-        #region Indentation management
-        public const string Indent = "    ";
-        private int _indentationLevel = 0;
-        private string _indentation = "";
-
-        public int IndentationLevel
-        {
-            get
-            {
-                return _indentationLevel;
-            }
-            set
-            {
-                _indentationLevel = value;
-
-                StringBuilder sb = new StringBuilder();
-                for (int i = 0; i < IndentationLevel; i++)
-                {
-                    sb.Append(Indent);
-                }
-                _indentation = sb.ToString();
-            }
-        }
-        public string Indentation => _indentation;
-        #endregion
-
-        #region Temp var management
-        /// <summary>
-        /// Maps a temp var to a place where it was created
-        /// </summary>
-        public Dictionary<string, TempVarAssignmentStatement> TempVarMap = new Dictionary<string, TempVarAssignmentStatement>();
-        /// <summary>
-        /// If used for auto-naming temp vars
-        /// </summary>
-        public int TempVarId { get; private set; }
-        public AssignmentStatement CompilerTempVar;
-
-        public TempVar NewTempVar()
-        {
-            return new TempVar(++TempVarId);
-        }
-        #endregion
-
-        #region Local var management
-        public HashSet<string> LocalVarDefines = new HashSet<string>();
-        #endregion
-
-        #region GMS 2.3+ Function management
-        /// <summary>
-        /// Set containing already-decompiled child code entries.
-        /// Used to prevent decompiling the same child entry multiple times.
-        /// Only applies to function entries, struct and constructors are unaffected.
-        /// </summary>
-        public ISet<UndertaleCode> AlreadyProcessed = new HashSet<UndertaleCode>();
-        #endregion
-
-        #region Asset type resolution
-        /// <summary>
-        /// Contains the resolved asset type for every variable
-        /// </summary>
-        public Dictionary<UndertaleVariable, AssetIDType> assetTypes = new Dictionary<UndertaleVariable, AssetIDType>();
-        public DirectFunctionCall currentFunction; // TODO: clean up this hack
-        #endregion
-
-        #region Decompilation results
-        /// <summary>
-        /// Contains the result of decompiling this code block.
-        /// This is a map from an entry point address to a list of statements.
-        /// Needs to be here to access it in ToString for inline function definitions.
-        /// </summary>
-        public Dictionary<uint, List<Statement>> Statements { get; internal set; }
-        #endregion
-
-        /// <summary>
-        /// Allows to disable the anonymous code name resolution to prevent recursion
-        /// </summary>
-        public bool DisableAnonymousFunctionNameResolution = false;
-    }
-
-    public static class Decompiler
+    public static partial class Decompiler
     {
         // Color dictionary for resolving.
         public static readonly Dictionary<uint, string> ColorDictionary = new Dictionary<uint, string>
@@ -212,1845 +37,7 @@ namespace UndertaleModLib.Decompiler
             [65535] = "c_yellow",
             [4235519] = "c_orange"
         };
-
-        // Represents a block node of instructions from GML bytecode (for control flow).
-        public class Block
-        {
-            public uint? Address;
-            public List<UndertaleInstruction> Instructions = new List<UndertaleInstruction>();
-            public List<Statement> Statements = null;
-            public Expression ConditionStatement = null;
-            public bool conditionalExit;
-            public Block nextBlockTrue;
-            public Block nextBlockFalse;
-            public List<Block> entryPoints = new List<Block>();
-            internal List<TempVarReference> TempVarsOnEntry;
-
-            public int _CachedIndex;
-
-            public Block(uint? address)
-            {
-                Address = address;
-            }
-
-            public override string ToString()
-            {
-                return "Block " + Address;
-            }
-        }
-
-        // Represents all kinds of high-level decompilation results, to be stringified at the end.
-        public abstract class Statement
-        {
-            public abstract string ToString(DecompileContext context);
-            internal abstract AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType);
-            public abstract Statement CleanStatement(DecompileContext context, BlockHLStatement block);
-        }
-
-        // Represents all expressions.
-        public abstract class Expression : Statement
-        {
-            public UndertaleInstruction.DataType Type;
-            public bool WasDuplicated = false;
-
-            // Helper function to convert opcode operations to "printable" strings.
-            public static string OperationToPrintableString(UndertaleInstruction.Opcode op)
-            {
-                return op switch
-                {
-                    UndertaleInstruction.Opcode.Mul => "*",
-                    UndertaleInstruction.Opcode.Div => "/",
-                    UndertaleInstruction.Opcode.Rem => "div",
-                    UndertaleInstruction.Opcode.Mod => "%",
-                    UndertaleInstruction.Opcode.Add => "+",
-                    UndertaleInstruction.Opcode.Sub => "-",
-                    UndertaleInstruction.Opcode.And => "&",
-                    UndertaleInstruction.Opcode.Or => "|",
-                    UndertaleInstruction.Opcode.Xor => "^",
-                    UndertaleInstruction.Opcode.Neg => "-",
-                    UndertaleInstruction.Opcode.Not => "~",
-                    UndertaleInstruction.Opcode.Shl => "<<",
-                    UndertaleInstruction.Opcode.Shr => ">>",
-                    _ => op.ToString().ToUpper(CultureInfo.InvariantCulture),
-                };
-            }
-
-            // Helper function to convert opcode comparisons to "printable" strings.
-            public static string OperationToPrintableString(UndertaleInstruction.ComparisonType op)
-            {
-                return op switch
-                {
-                    UndertaleInstruction.ComparisonType.LT => "<",
-                    UndertaleInstruction.ComparisonType.LTE => "<=",
-                    UndertaleInstruction.ComparisonType.EQ => "==",
-                    UndertaleInstruction.ComparisonType.NEQ => "!=",
-                    UndertaleInstruction.ComparisonType.GTE => ">=",
-                    UndertaleInstruction.ComparisonType.GT => ">",
-                    _ => op.ToString().ToUpper(CultureInfo.InvariantCulture),
-                };
-            }
-
-            internal virtual bool IsDuplicationSafe()
-            {
-                return false;
-            }
-
-            public Expression CleanExpression(DecompileContext context, BlockHLStatement block)
-            {
-                return CleanStatement(context, block) as Expression;
-            }
-        }
-
-        // Represents a base constant expression, such as a number (no operations).
-        public class ExpressionConstant : Expression
-        {
-            public object Value;
-            public bool IsPushE;
-            internal AssetIDType AssetType = AssetIDType.Other;
-
-            public ExpressionConstant(UndertaleInstruction.DataType type, object value, bool isPushE = false)
-            {
-                Type = type;
-                Value = value;
-                IsPushE = isPushE;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return true;
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Value = (Value as Statement)?.CleanStatement(context, block) ?? Value;
-                return this;
-            }
-
-            public bool EqualsNumber(int TestNumber)
-            {
-                return (Value is Int16 || Value is Int32) && Convert.ToInt32(Value) == TestNumber;
-            }
-
-            // Helper function to carefully check if an object is in fact an integer, for asset types.
-            public static int? ConvertToInt(object val)
-            {
-                if (val is int || val is short || val is ushort || val is UndertaleInstruction.InstanceType)
-                {
-                    return Convert.ToInt32(val);
-                }
-                else if (val is double)
-                {
-                    var v = Convert.ToDouble(val);
-                    int res = (int)v;
-                    if (v == res)
-                        return res;
-                }
-                else if (val is float)
-                {
-                    var v = Convert.ToSingle(val);
-                    int res = (int)v;
-                    if (v == res)
-                        return res;
-                }
-                return null;
-            }
-
-            // Helper function, using the one above, to convert an object into its respective asset type enum, if possible.
-            public static string ConvertToEnumStr<T>(object val)
-            {
-                int? intVal = ConvertToInt(val);
-                if (intVal == null)
-                    return val.ToString();
-                return ((T)(object)intVal).ToString();
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                if (Value is float f) // More accurate, larger range, double to string.
-                    return RoundTrip.ToRoundTrip(f);
-
-                if (Value is Int64 i && i <= int.MaxValue && i >= int.MinValue) // Decompiler accuracy improvement.
-                {
-                    return "(" + i + " << 0)";
-                }
-
-                if (Value is double d) // More accurate, larger range, double to string.
-                    return RoundTrip.ToRoundTrip(d);
-
-                if (Value is Statement statement)
-                    return statement.ToString(context);
-
-                if (Value is UndertaleResourceById<UndertaleString, UndertaleChunkSTRG> resource) // Export string.
-                {
-                    string resultStr = resource.Resource.ToString(context.GlobalContext.Data?.IsGameMaker2() ?? false);
-                    if (context.GlobalContext.EnableStringLabels)
-                        resultStr += resource.GetMarkerSuffix();
-                    return resultStr;
-                }
-
-                // Archie: If statements are inefficient! Use a switch jump table!
-                if (AssetType == AssetIDType.GameObject && !(Value is Int64)) // When the value is Int64, an example value is 343434343434. It is unknown what it represents, but it's not an InstanceType.
-                {
-                    int? val = ConvertToInt(Value);
-                    if (val != null && val < 0 && val >= -16)
-                        return ((UndertaleInstruction.InstanceType)Value).ToString().ToLower(CultureInfo.InvariantCulture);
-                }
-                else switch (AssetType) // Need to put else because otherwise it gets terribly unoptimized with GameObject type
-                {
-                    case AssetIDType.e__VW:
-                        return "e__VW." + ConvertToEnumStr<e__VW>(Value);
-                    case AssetIDType.e__BG:
-                        return "e__BG." + ConvertToEnumStr<e__BG>(Value);
-
-                    case AssetIDType.Enum_HAlign:
-                        return ConvertToEnumStr<HAlign>(Value);
-                    case AssetIDType.Enum_VAlign:
-                        return ConvertToEnumStr<VAlign>(Value);
-                    case AssetIDType.Enum_GameSpeed:
-                        return ConvertToEnumStr<GameSpeed>(Value);
-                    case AssetIDType.Enum_OSType:
-                        return ConvertToEnumStr<OSType>(Value);
-                    case AssetIDType.Enum_GamepadButton:
-                        return ConvertToEnumStr<GamepadButton>(Value);
-                    case AssetIDType.MouseButton:
-                        return ConvertToEnumStr<MouseButton>(Value);
-                    case AssetIDType.Enum_MouseCursor:
-                        return ConvertToEnumStr<MouseCursor>(Value);
-                    case AssetIDType.Enum_PathEndAction:
-                        return ConvertToEnumStr<PathEndAction>(Value);
-                    case AssetIDType.Enum_BufferKind:
-                        return ConvertToEnumStr<BufferKind>(Value);
-                    case AssetIDType.Enum_BufferType:
-                        return ConvertToEnumStr<BufferType>(Value);
-                    case AssetIDType.Enum_BufferSeek:
-                        return ConvertToEnumStr<BufferSeek>(Value);
-                    case AssetIDType.Enum_Steam_UGC_FileType:
-                        return ConvertToEnumStr<Steam_UGC_FileType>(Value);
-                    case AssetIDType.Enum_Steam_UGC_List:
-                        return ConvertToEnumStr<Steam_UGC_List>(Value);
-                    case AssetIDType.Enum_Steam_UGC_MatchType:
-                        return ConvertToEnumStr<Steam_UGC_MatchType>(Value);
-                    case AssetIDType.Enum_Steam_UGC_QueryType:
-                        return ConvertToEnumStr<Steam_UGC_QueryType>(Value);
-                    case AssetIDType.Enum_Steam_UGC_SortOrder:
-                        return ConvertToEnumStr<Steam_UGC_SortOrder>(Value);
-                    case AssetIDType.Enum_Steam_Overlay:
-                        return ConvertToEnumStr<Steam_Overlay>(Value);
-                    case AssetIDType.Enum_Steam_LeaderBoard_Display:
-                        return ConvertToEnumStr<Steam_LeaderBoard_Display>(Value);
-                    case AssetIDType.Enum_Steam_LeaderBoard_Sort:
-                        return ConvertToEnumStr<Steam_LeaderBoard_Sort>(Value);
-                    case AssetIDType.Boolean:
-                        return ConvertToEnumStr<Boolean>(Value);
-                    case AssetIDType.EventType:
-                        return ConvertToEnumStr<Enum_EventType>(Value);
-                    case AssetIDType.ContextDependent:
-                        {
-                            var func = context.currentFunction;
-                            if (func != null && (ContextualAssetResolver.resolvers?.ContainsKey(func.Function.Name.Content) ?? false))
-                            {
-                                List<Expression> actualArguments = new List<Expression>();
-                                foreach (var arg in func.Arguments)
-                                {
-                                    if (arg is ExpressionCast)
-                                        actualArguments.Add((arg as ExpressionCast).Argument);
-                                    else
-                                        actualArguments.Add(arg);
-                                }
-                                string result = ContextualAssetResolver.resolvers[func.Function.Name.Content](context, func, actualArguments.IndexOf(this), this);
-                                if (result != null)
-                                    return result;
-                            }
-                        }
-                        break;
-
-                    case AssetIDType.Color:
-                        if (Value is IFormattable formattable && !(Value is float) && !(Value is double) && !(Value is decimal))
-                        {
-                            int vint = Convert.ToInt32(Value);
-                            if (vint < 0) // negative value.
-                                return vint.ToString();
-                            else // guaranteed to be an unsigned int.
-                            {
-                                uint vuint = (uint)vint;
-                                if (Decompiler.ColorDictionary.ContainsKey(vuint))
-                                    return Decompiler.ColorDictionary[vuint];
-                                else
-                                    return (context.GlobalContext.Data?.IsGameMaker2() ?? false ? "0x" : "$") + formattable.ToString("X6", CultureInfo.InvariantCulture); // not a known color and not negative.
-                            }
-                        }
-                        break;
-
-                    case AssetIDType.KeyboardKey:
-                        {
-                            string key = GetAsKeyboard(context);
-                            if (key != null)
-                                return key;
-                        }
-                        break;
-                    // Don't use this.
-                    // It will not recompile.
-                    case AssetIDType.Macro:
-                        throw new NotImplementedException();/*
-                        {
-                            var macros = ContextualAssetResolver.macros;
-                            var key = Value?.ToString();
-
-                            if (key != null & macros.ContainsKey(key))
-                                return macros[key];
-                        }
-                        break;*/
-                }
-
-                if ((context.AssetResolutionEnabled || AssetType == AssetIDType.Script) && context.GlobalContext.Data != null && AssetType != AssetIDType.Other)
-                {
-                    IList assetList = null;
-                    switch (AssetType)
-                    {
-                        case AssetIDType.Sprite:
-                            assetList = (IList)context.GlobalContext.Data.Sprites;
-                            break;
-                        case AssetIDType.TileSet:
-                        case AssetIDType.Background:
-                            assetList = (IList)context.GlobalContext.Data.Backgrounds;
-                            break;
-                        case AssetIDType.Sound:
-                            assetList = (IList)context.GlobalContext.Data.Sounds;
-                            break;
-                        case AssetIDType.Font:
-                            assetList = (IList)context.GlobalContext.Data.Fonts;
-                            break;
-                        case AssetIDType.Path:
-                            assetList = (IList)context.GlobalContext.Data.Paths;
-                            break;
-                        case AssetIDType.Timeline:
-                            assetList = (IList)context.GlobalContext.Data.Timelines;
-                            break;
-                        case AssetIDType.Room:
-                            assetList = (IList)context.GlobalContext.Data.Rooms;
-                            break;
-                        case AssetIDType.GameObject:
-                            assetList = (IList)context.GlobalContext.Data.GameObjects;
-                            break;
-                        case AssetIDType.Shader:
-                            assetList = (IList)context.GlobalContext.Data.Shaders;
-                            break;
-                        case AssetIDType.Script:
-                            assetList = (IList)context.GlobalContext.Data.Scripts;
-                            break;
-                    }
-
-                    if (!(Value is Int64)) // It is unknown what Int64 data represents, but it's not this.
-                    {
-                        int? tryVal = ConvertToInt(Value);
-                        int val;
-                        if (tryVal != null)
-                        {
-                            val = tryVal ?? -1;
-                            if (assetList != null && val >= 0 && val < assetList.Count)
-                                return ((UndertaleNamedResource)assetList[val]).Name.Content;
-                        }
-                    }
-                }
-
-                return ((Value as IFormattable)?.ToString(null, CultureInfo.InvariantCulture) ?? Value.ToString());
-            }
-
-            // Helper function
-            public string GetAsKeyboard(DecompileContext context)
-            {
-                int? tryVal = ConvertToInt(Value);
-                if (tryVal != null)
-                {
-                    int val = tryVal ?? -1;
-
-                    bool isAlphaNumeric = val >= (int)EventSubtypeKey.Digit0 && val <= (int)EventSubtypeKey.Z;
-                    if (isAlphaNumeric)
-                        return "ord(\"" + (char)val + "\")";
-
-                    if (val >= 0 && Enum.IsDefined(typeof(EventSubtypeKey), (uint)val))
-                        return ((EventSubtypeKey)val).ToString(); // Either return the key enum, or the right alpha-numeric key-press.
-
-                    if (!Char.IsControl((char)val) && !Char.IsLower((char)val) && val > 0) // The special keys overlay with the uppercase letters (ugh)
-                        return "ord(" + (((char)val) == '\'' ? (context.GlobalContext.Data?.IsGameMaker2() ?? false ? "\"\\\"\"" : "'\"'")
-                            : (((char)val) == '\\' ? (context.GlobalContext.Data?.IsGameMaker2() ?? false ? "\"\\\\\"" : "\"\\\"")
-                            : "\"" + (char)val + "\"")) + ")";
-                }
-                return null;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                if (AssetType == AssetIDType.Other)
-                    AssetType = suggestedType;
-                return AssetType;
-            }
-        }
-
-        // Represents a reference to an asset in the resource tree, used in 2023.8+ only
-        public class ExpressionAssetRef : Expression
-        {
-            // NOTE: Also see generalized "ResourceType" enum. This has slightly differing values, though
-            public enum RefType
-            {
-                Object = 0,
-                Sprite = 1,
-                Sound = 2,
-                Room = 3,
-                Background = 4,
-                Path = 5,
-                // missing 6
-                Font = 7,
-                Timeline = 8,
-                // missing 9
-                Shader = 10,
-                Sequence = 11,
-                AnimCurve = 12,
-                ParticleSystem = 13
-            }
-
-            public int AssetIndex;
-            public RefType AssetRefType;
-
-            public ExpressionAssetRef(int encodedResourceIndex)
-            {
-                Type = UndertaleInstruction.DataType.Variable;
-
-                // Break down index - first 24 bits are the ID, the rest is the ref type
-                AssetIndex = encodedResourceIndex & 0xffffff;
-                AssetRefType = (RefType)(encodedResourceIndex >> 24);
-            }
-
-            public ExpressionAssetRef(int resourceIndex, RefType resourceType)
-            {
-                Type = UndertaleInstruction.DataType.Variable;
-                AssetIndex = resourceIndex;
-                AssetRefType = resourceType;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return true;
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                return this;
-            }
-            public override string ToString(DecompileContext context)
-            {
-                if (context.GlobalContext.Data != null)
-                {
-                    IList assetList = null;
-                    switch (AssetRefType)
-                    {
-                        case RefType.Sprite:
-                            assetList = (IList)context.GlobalContext.Data.Sprites;
-                            break;
-                        case RefType.Background:
-                            assetList = (IList)context.GlobalContext.Data.Backgrounds;
-                            break;
-                        case RefType.Sound:
-                            assetList = (IList)context.GlobalContext.Data.Sounds;
-                            break;
-                        case RefType.Font:
-                            assetList = (IList)context.GlobalContext.Data.Fonts;
-                            break;
-                        case RefType.Path:
-                            assetList = (IList)context.GlobalContext.Data.Paths;
-                            break;
-                        case RefType.Timeline:
-                            assetList = (IList)context.GlobalContext.Data.Timelines;
-                            break;
-                        case RefType.Room:
-                            assetList = (IList)context.GlobalContext.Data.Rooms;
-                            break;
-                        case RefType.Object:
-                            assetList = (IList)context.GlobalContext.Data.GameObjects;
-                            break;
-                        case RefType.Shader:
-                            assetList = (IList)context.GlobalContext.Data.Shaders;
-                            break;
-                        case RefType.AnimCurve:
-                            assetList = (IList)context.GlobalContext.Data.AnimationCurves;
-                            break;
-                        case RefType.Sequence:
-                            assetList = (IList)context.GlobalContext.Data.Sequences;
-                            break;
-                        case RefType.ParticleSystem:
-                            assetList = (IList)context.GlobalContext.Data.ParticleSystems;
-                            break;
-                    }
-
-                    if (assetList != null && AssetIndex >= 0 && AssetIndex < assetList.Count)
-                        return ((UndertaleNamedResource)assetList[AssetIndex]).Name.Content;
-                }
-                return $"/* ERROR: missing {AssetRefType} asset, using ID instead */ {AssetIndex}";
-            }
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                // Convert type to corresponding AssetIDType equivalent
-                return AssetRefType switch
-                {
-                    RefType.Object => AssetIDType.GameObject,
-                    RefType.Sprite => AssetIDType.Sprite,
-                    RefType.Sound => AssetIDType.Sound,
-                    RefType.Room => AssetIDType.Room,
-                    RefType.Background => AssetIDType.Background,
-                    RefType.Path => AssetIDType.Path,
-                    RefType.Font => AssetIDType.Font,
-                    RefType.Timeline => AssetIDType.Timeline,
-                    RefType.Shader => AssetIDType.Shader,
-                    RefType.Sequence => AssetIDType.Sequence,
-                    RefType.AnimCurve => AssetIDType.AnimCurve,
-                    RefType.ParticleSystem => AssetIDType.ParticleSystem,
-                    _ => throw new NotImplementedException($"Missing ref type {AssetRefType}")
-                };
-            }
-        }
-
-        // Represents an expression converted to one of another data type - makes no difference on high-level code.
-        public class ExpressionCast : Expression
-        {
-            public Expression Argument;
-
-            public ExpressionCast(UndertaleInstruction.DataType targetType, Expression argument)
-            {
-                this.Type = targetType;
-                this.Argument = argument;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return Argument.IsDuplicationSafe();
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Argument = Argument?.CleanExpression(context, block);
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                return Argument.ToString(context);
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return Argument.DoTypePropagation(context, suggestedType);
-            }
-        }
-
-        // Represents a unary expression.
-        public class ExpressionOne : Expression
-        {
-            public UndertaleInstruction.Opcode Opcode;
-            public Expression Argument;
-
-            public ExpressionOne(UndertaleInstruction.Opcode opcode, UndertaleInstruction.DataType targetType, Expression argument)
-            {
-                this.Opcode = opcode;
-                this.Type = targetType;
-                this.Argument = argument;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return Argument.IsDuplicationSafe();
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Argument = Argument?.CleanExpression(context, block);
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                string op = OperationToPrintableString(Opcode);
-                if (Opcode == UndertaleInstruction.Opcode.Not && Type == UndertaleInstruction.DataType.Boolean)
-                    op = "!"; // This is a logical negation instead, see #93
-                string arg = Argument.ToString(context);
-                if (arg.Contains(' ', StringComparison.InvariantCulture))
-                    return String.Format("({0}({1}))", op, arg);
-                return String.Format("({0}{1})", op, arg);
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return Argument.DoTypePropagation(context, suggestedType);
-            }
-        }
-
-        // This is basically ExpressionTwo, but allows for using symbols like && or || without creating new opcodes.
-        public class ExpressionTernary : Expression
-        {
-            public Expression Condition;
-            public Expression TrueExpression;
-            public Expression FalseExpression;
-
-            public ExpressionTernary(UndertaleInstruction.DataType targetType, Expression Condition, Expression argument1, Expression argument2)
-            {
-                this.Type = targetType;
-                this.Condition = Condition;
-                this.TrueExpression = argument1;
-                this.FalseExpression = argument2;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return Condition.IsDuplicationSafe() && TrueExpression.IsDuplicationSafe() && FalseExpression.IsDuplicationSafe();
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Condition = Condition?.CleanExpression(context, block);
-                TrueExpression = TrueExpression?.CleanExpression(context, block);
-                FalseExpression = FalseExpression?.CleanExpression(context, block);
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                string condStr = Condition.ToString(context);
-                if (TestNumber(TrueExpression, 1) && TestNumber(FalseExpression, 0))
-                    return condStr; // Default values, yes = true, no = false.
-
-                return "(" + condStr + " ? " + TrueExpression.ToString(context) + " : " + FalseExpression.ToString(context) + ")";
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                // The most likely, but probably rarely happens
-                AssetIDType t = TrueExpression.DoTypePropagation(context, suggestedType);
-                FalseExpression.DoTypePropagation(context, AssetIDType.Other);
-                return t;
-            }
-        }
-
-        // Represents post increments and decrements, such as a++ and a--.
-        public class ExpressionPost : Expression
-        {
-            public UndertaleInstruction.Opcode Opcode;
-            public Expression Variable;
-
-            public ExpressionPost(UndertaleInstruction.Opcode opcode, Expression variable)
-            {
-                Opcode = opcode;
-                Variable = variable;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return Variable.IsDuplicationSafe();
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Variable = Variable?.CleanStatement(context, block) as Expression;
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                return Variable.ToString(context) + (Opcode == UndertaleInstruction.Opcode.Add ? "++" : "--");
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return Variable.DoTypePropagation(context, suggestedType);
-            }
-        }
-
-        // Represents pre increments and decrements, such as ++a and --a.
-        public class ExpressionPre : Expression
-        {
-            public UndertaleInstruction.Opcode Opcode;
-            public Expression Variable;
-
-            public ExpressionPre(UndertaleInstruction.Opcode opcode, Expression variable)
-            {
-                Opcode = opcode;
-                Variable = variable;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return Variable.IsDuplicationSafe();
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Variable = Variable?.CleanExpression(context, block);
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                return (Opcode == UndertaleInstruction.Opcode.Add ? "++" : "--") + Variable.ToString(context);
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return Variable.DoTypePropagation(context, suggestedType);
-            }
-        }
-
-        // This is basically ExpressionTwo, but allows for using symbols like && or || without creating new opcodes.
-        public class ExpressionTwoSymbol : Expression
-        {
-            public string Symbol;
-            public Expression Argument1;
-            public Expression Argument2;
-
-            public ExpressionTwoSymbol(string symbol, UndertaleInstruction.DataType targetType, Expression argument1, Expression argument2)
-            {
-                this.Symbol = symbol;
-                this.Type = targetType;
-                this.Argument1 = argument1;
-                this.Argument2 = argument2;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return Argument1.IsDuplicationSafe() && Argument2.IsDuplicationSafe();
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Argument1 = Argument1?.CleanExpression(context, block);
-                Argument2 = Argument2?.CleanExpression(context, block);
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                string arg1;
-                if (Argument1 is ExpressionTwoSymbol && (Argument1 as ExpressionTwoSymbol).Symbol == Symbol)
-                    arg1 = (Argument1 as ExpressionTwoSymbol).ToStringNoParen(context);
-                else
-                    arg1 = Argument1.ToString(context);
-                string arg2;
-                if (Argument2 is ExpressionTwoSymbol && (Argument2 as ExpressionTwoSymbol).Symbol == Symbol)
-                    arg2 = (Argument2 as ExpressionTwoSymbol).ToStringNoParen(context);
-                else
-                    arg2 = Argument2.ToString(context);
-                return String.Format("({0} {1} {2})", arg1, Symbol, arg2);
-            }
-
-            public string ToStringNoParen(DecompileContext context)
-            {
-                string arg1;
-                if (Argument1 is ExpressionTwoSymbol && (Argument1 as ExpressionTwoSymbol).Symbol == Symbol)
-                    arg1 = (Argument1 as ExpressionTwoSymbol).ToStringNoParen(context);
-                else
-                    arg1 = Argument1.ToString(context);
-                string arg2;
-                if (Argument2 is ExpressionTwoSymbol && (Argument2 as ExpressionTwoSymbol).Symbol == Symbol)
-                    arg2 = (Argument2 as ExpressionTwoSymbol).ToStringNoParen(context);
-                else
-                    arg2 = Argument2.ToString(context);
-                return String.Format("{0} {1} {2}", arg1, Symbol, arg2);
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                // The most likely, but probably rarely happens
-                AssetIDType t = Argument1.DoTypePropagation(context, suggestedType);
-                Argument2.DoTypePropagation(context, AssetIDType.Other);
-                return t;
-            }
-        }
-
-        // Represents a binary expression.
-        public class ExpressionTwo : Expression
-        {
-            public UndertaleInstruction.Opcode Opcode;
-            public UndertaleInstruction.DataType Type2;
-            public Expression Argument1;
-            public Expression Argument2;
-
-            public ExpressionTwo(UndertaleInstruction.Opcode opcode, UndertaleInstruction.DataType targetType, UndertaleInstruction.DataType type2, Expression argument1, Expression argument2)
-            {
-                this.Opcode = opcode;
-                this.Type = targetType;
-                this.Type2 = type2;
-                this.Argument1 = argument1;
-                this.Argument2 = argument2;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return Argument1.IsDuplicationSafe() && Argument2.IsDuplicationSafe();
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Argument1 = Argument1?.CleanExpression(context, block);
-                Argument2 = Argument2?.CleanExpression(context, block);
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                if (Opcode == UndertaleInstruction.Opcode.Or || Opcode == UndertaleInstruction.Opcode.And)
-                {
-                    // If both arguments are a boolean type, this is a non-short-circuited logical condition
-                    if (Type == UndertaleInstruction.DataType.Boolean && Type2 == UndertaleInstruction.DataType.Boolean)
-                        return String.Format("({0} {1}{1} {2})", Argument1.ToString(context), OperationToPrintableString(Opcode), Argument2.ToString(context));
-                }
-                return String.Format("({0} {1} {2})", Argument1.ToString(context), OperationToPrintableString(Opcode), Argument2.ToString(context));
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                // The most likely, but probably rarely happens
-                AssetIDType t = Argument1.DoTypePropagation(context, suggestedType);
-                Argument2.DoTypePropagation(context, AssetIDType.Other);
-                return t;
-            }
-        }
-
-        // Represents a binary comparison expression.
-        public class ExpressionCompare : Expression
-        {
-            public UndertaleInstruction.ComparisonType Opcode;
-            public Expression Argument1;
-            public Expression Argument2;
-
-            public ExpressionCompare(UndertaleInstruction.ComparisonType opcode, Expression argument1, Expression argument2)
-            {
-                this.Opcode = opcode;
-                this.Type = UndertaleInstruction.DataType.Boolean;
-                this.Argument1 = argument1;
-                this.Argument2 = argument2;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return Argument1.IsDuplicationSafe() && Argument2.IsDuplicationSafe();
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                string arg1, arg2;
-                if (Argument1 is ExpressionCompare)
-                    arg1 = (Argument1 as ExpressionCompare).ToStringWithParen(context);
-                else
-                    arg1 = Argument1.ToString(context);
-                if (Argument2 is ExpressionCompare)
-                    arg2 = (Argument2 as ExpressionCompare).ToStringWithParen(context);
-                else
-                    arg2 = Argument2.ToString(context);
-                return String.Format("{0} {1} {2}", arg1, OperationToPrintableString(Opcode), arg2);
-            }
-
-            public string ToStringWithParen(DecompileContext context)
-            {
-                return "(" + ToString(context) + ")";
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Argument1 = Argument1?.CleanExpression(context, block);
-                Argument2 = Argument2?.CleanExpression(context, block);
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                // TODO: This should be probably able to go both ways...
-                Argument2.DoTypePropagation(context, Argument1.DoTypePropagation(context, suggestedType));
-
-                /*
-                if (Opcode != UndertaleInstruction.ComparisonType.EQ && Opcode != UndertaleInstruction.ComparisonType.NEQ)
-                {
-                    if (Argument1 is ExpressionConstant arg1)
-                        if (arg1.AssetType == AssetIDType.Script)
-                            arg1.AssetType = AssetIDType.Other;
-
-                    if (Argument2 is ExpressionConstant arg2)
-                        if (arg2.AssetType == AssetIDType.Script)
-                            arg2.AssetType = AssetIDType.Other;
-                }*/
-
-                return AssetIDType.Other;
-            }
-        }
-
-        // Represents an unnamed value that gets passed around the stack.
-        // Theoretically, these should be cleaned up and removed by the end of decompilation.
-        public class TempVar
-        {
-            public string Name;
-            public UndertaleInstruction.DataType Type;
-            internal AssetIDType AssetType;
-
-            public TempVar(int id)
-            {
-                Name = MakeTemporaryVarName(id);
-            }
-
-            public static string MakeTemporaryVarName(int id)
-            {
-                return "_temp_local_var_" + id;
-            }
-        }
-
-        // A reference class for tempvars.
-        public class TempVarReference
-        {
-            public TempVar Var;
-
-            public TempVarReference(TempVar var)
-            {
-                Var = var;
-            }
-        }
-
-        // Assignment statement for tempvars.
-        public class TempVarAssignmentStatement : Statement
-        {
-            public TempVarReference Var;
-            public Expression Value;
-
-            public bool HasVarKeyword;
-
-            public TempVarAssignmentStatement(TempVarReference var, Expression value)
-            {
-                Var = var;
-                Value = value;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                //TODO: why is there a GMS2Check for this? var exists in gms1.4 as well
-                if (context.GlobalContext.Data?.IsGameMaker2() ?? false && !HasVarKeyword && context.LocalVarDefines.Add(Var.Var.Name))
-                    HasVarKeyword = true;
-
-                return String.Format("{0}{1} = {2}", (HasVarKeyword ? "var " : ""), Var.Var.Name, Value.ToString(context));
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Value = Value?.CleanExpression(context, block);
-
-                if (Var != null && Var.Var != null && Var.Var.Name != null)
-                {
-                    if ((Value as ExpressionTempVar)?.Var?.Var?.Name == Var.Var.Name) // This is literally set to itself. No thank you.
-                    {
-                        block.Statements.Remove(context.TempVarMap[Var.Var.Name]);
-                        return context.TempVarMap[Var.Var.Name].CleanStatement(context, block);
-                    }
-
-                    context.TempVarMap[Var.Var.Name] = this;
-                }
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                if (Var.Var.AssetType == AssetIDType.Other)
-                    Var.Var.AssetType = suggestedType;
-                return Value.DoTypePropagation(context, Var.Var.AssetType);
-            }
-        }
-
-        // Represents a tempvar inside of an expression.
-        public class ExpressionTempVar : Expression
-        {
-            public TempVarReference Var;
-
-            public ExpressionTempVar(TempVarReference var, UndertaleInstruction.DataType targetType)
-            {
-                this.Var = var;
-                this.Type = targetType;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                return true;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                return Var.Var.Name;
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                TempVarAssignmentStatement tempVarStatement = context.TempVarMap[Var.Var.Name];
-                if (tempVarStatement != null)
-                {
-                    block.Statements.Remove(tempVarStatement);
-                    return tempVarStatement.Value.CleanStatement(context, block);
-                }
-
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                if (Var.Var.AssetType == AssetIDType.Other)
-                    Var.Var.AssetType = suggestedType;
-                return Var.Var.AssetType;
-            }
-        }
-
-        // Represents a high-level return statement, or an exit in Studio version < 2 if there is no value.
-        public class ReturnStatement : Statement
-        {
-            public Expression Value;
-
-            public ReturnStatement(Expression value)
-            {
-                Value = value;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                if (Value != null)
-                {
-                    if (AssetTypeResolver.return_types.ContainsKey(context.TargetCode.Name.Content))
-                        Value.DoTypePropagation(context, AssetTypeResolver.return_types[context.TargetCode.Name.Content]);
-                    if (context.GlobalContext.Data != null && !DecompileContext.GMS2_3)
-                    {
-                        // We might be decompiling a legacy script - resolve it's name
-                        UndertaleScript script = context.GlobalContext.Data.Scripts.FirstOrDefault(x => x.Code == context.TargetCode);
-                        if (script != null && AssetTypeResolver.return_types.ContainsKey(script.Name.Content))
-                            Value.DoTypePropagation(context, AssetTypeResolver.return_types[script.Name.Content]);
-                    }
-
-                    string cleanVal = Value.ToString(context);
-                    if (cleanVal.EndsWith("\n", StringComparison.InvariantCulture))
-                        cleanVal = cleanVal.Substring(0, cleanVal.Length - 1);
-
-                    return "return " + cleanVal + ";";
-                }
-                else
-                    return (context.GlobalContext.Data?.IsGameMaker2() ?? false ? "return;" : "exit");
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Value = Value?.CleanExpression(context, block);
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return Value?.DoTypePropagation(context, suggestedType) ?? suggestedType;
-            }
-        }
-
-        // Represents a high-level assignment statement.
-        public class AssignmentStatement : Statement
-        {
-            public ExpressionVar Destination;
-            public Expression Value;
-
-            public bool HasVarKeyword;
-
-            private bool _isStructDefinition, _checkedForDefinition;
-            public bool IsStructDefinition
-            {
-                get
-                {
-                    // Quick hack
-                    if (!_checkedForDefinition)
-                    {
-                        try
-                        {
-                            if (Destination.Var.Name.Content.StartsWith("___struct___", StringComparison.InvariantCulture))
-                            {
-                                Expression val = Value;
-                                while (val is ExpressionCast cast)
-                                    val = cast;
-
-                                if (val is FunctionDefinition def)
-                                {
-                                    def.PromoteToStruct();
-                                    _isStructDefinition = true;
-                                }
-                            }
-                        }
-                        catch (Exception) { }
-                        _checkedForDefinition = true;
-                    }
-                    return _isStructDefinition;
-                }
-            }
-
-            public AssignmentStatement(ExpressionVar destination, Expression value)
-            {
-                Destination = destination;
-                Value = value;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                bool gms2 = context.GlobalContext.Data?.IsGameMaker2() ?? false;
-
-                if (gms2 && IsStructDefinition)
-                    return "";
-
-                string varName = Destination.ToString(context);
-
-                if (gms2 && !HasVarKeyword)
-                {
-                    var data = context.GlobalContext.Data;
-                    if (data != null)
-                    {
-                        var locals = data.CodeLocals.For(context.TargetCode);
-                        // Stop decompiler from erroring on missing CodeLocals
-                        if (locals != null && locals.HasLocal(varName) && context.LocalVarDefines.Add(varName))
-                            HasVarKeyword = true;
-                    }
-                }
-
-                // Someone enlighten me on structs, I'm steering clear for now.
-                // And find the "right" way to do this.
-                if (Value is FunctionDefinition functionVal && functionVal.Subtype != FunctionDefinition.FunctionType.Struct)
-                {
-                    functionVal.IsStatement = true;
-                    return functionVal.ToString(context);
-                }
-
-                string varPrefix = (HasVarKeyword ? "var " : "");
-
-                // Check for possible ++, --, or operation equal (for single vars)
-                if (Value is ExpressionTwo two && (two.Argument1 is ExpressionVar) &&
-                    (two.Argument1 as ExpressionVar).Var == Destination.Var)
-                {
-                    if (two.Argument2 is ExpressionConstant c && c.IsPushE && ExpressionConstant.ConvertToInt(c.Value) == 1)
-                        return varName + (two.Opcode == UndertaleInstruction.Opcode.Add ? "++" : "--");
-
-                    // Not ++ or --, could potentially be an operation equal
-                    bool checkEqual(ExpressionVar a, ExpressionVar b)
-                    {
-                        if (a.InstType.GetType() != b.InstType.GetType())
-                            return false;
-                        ExpressionConstant ac = (a.InstType as ExpressionConstant), bc = (b.InstType as ExpressionConstant);
-                        bool res = ac.Value.Equals(bc.Value) && ac.IsPushE == bc.IsPushE && ac.Type == bc.Type && ac.WasDuplicated == bc.WasDuplicated &&
-                               a.VarType == b.VarType;
-                        res &= (a.ArrayIndices != null) == (b.ArrayIndices != null);
-                        if (a.ArrayIndices != null)
-                        {
-                            res &= a.ArrayIndices.Count == b.ArrayIndices.Count;
-                            if (res)
-                            {
-                                for (int i = 0; i < a.ArrayIndices.Count; i++)
-                                    res &= a.ArrayIndices[i] == b.ArrayIndices[i];
-                            }
-                        }
-                        return res;
-                    }
-                    if (Destination.InstType is ExpressionConstant)
-                    {
-                        ExpressionVar v1 = (ExpressionVar)two.Argument1;
-                        if (checkEqual(Destination, v1) && two.Opcode != UndertaleInstruction.Opcode.Shl && two.Opcode != UndertaleInstruction.Opcode.Shr && two.Opcode != UndertaleInstruction.Opcode.Rem)
-                        {
-                            if (!(context.GlobalContext.Data?.GeneralInfo?.BytecodeVersion > 14 && v1.Opcode != UndertaleInstruction.Opcode.Push && Destination.Var.InstanceType != UndertaleInstruction.InstanceType.Self))
-                                return String.Format("{0}{1} {2}= {3}", varPrefix, varName, Expression.OperationToPrintableString(two.Opcode), two.Argument2.ToString(context));
-                        }
-                    }
-                }
-                return String.Format("{0}{1}{2} {3}", varPrefix, varName, context.DecompilingStruct ? ":" : " =", Value.ToString(context));
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Expression expr = Destination.CleanExpression(context, block);
-
-                if (expr is ExpressionVar expvar)
-                    Destination = expvar;
-
-                Value = Value.CleanExpression(context, block);
-                if (Destination.Var.Name?.Content == "$$$$temp$$$$")
-                    context.CompilerTempVar = this;
-
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return Value.DoTypePropagation(context, Destination.DoTypePropagation(context, suggestedType));
-            }
-        }
-
-        // Represents a high-level operation-equals statement, such as a += 1.
-        public class OperationEqualsStatement : Statement
-        {
-            public ExpressionVar Destination;
-            public UndertaleInstruction.Opcode Operation;
-            public Expression Value;
-
-            public OperationEqualsStatement(ExpressionVar destination, UndertaleInstruction.Opcode operation, Expression value)
-            {
-                Destination = destination;
-                Operation = operation;
-                Value = value;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                return String.Format("{0} {1}= {2}", Destination.ToString(context), Expression.OperationToPrintableString(Operation), Value.ToString(context));
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                Value = Value?.CleanExpression(context, block);
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return Value.DoTypePropagation(context, Destination.DoTypePropagation(context, suggestedType));
-            }
-        }
-
-        // Represents a code comment, for debugging use (or minor error reporting).
-        public class CommentStatement : Statement
-        {
-            public string Message;
-
-            public CommentStatement(string message)
-            {
-                Message = message;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                return "// " + Message;
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return suggestedType;
-            }
-        }
-
-        // Represents an inline function definition
-        public class FunctionDefinition : Expression
-        {
-            public enum FunctionType
-            {
-                Function,
-                Constructor,
-                Struct
-            }
-
-            public UndertaleFunction Function { get; private set; }
-            public UndertaleCode FunctionBodyCodeEntry { get; private set; }
-            public Block FunctionBodyEntryBlock { get; private set; }
-            public FunctionType Subtype { get; private set; }
-            public bool IsStatement = false; // I know it's an expression, yes. But I'm not duplicating the rest.
-
-            internal List<Expression> Arguments;
-
-            public FunctionDefinition(UndertaleFunction target, UndertaleCode functionBodyCodeEntry, Block functionBodyEntryBlock, FunctionType type)
-            {
-                Subtype = type;
-                Function = target;
-                FunctionBodyCodeEntry = functionBodyCodeEntry;
-                FunctionBodyEntryBlock = functionBodyEntryBlock;
-            }
-
-            public void PromoteToStruct()
-            {
-                if (Subtype == FunctionType.Function)
-                    throw new InvalidOperationException("Cannot promote function to struct");
-
-                Subtype = FunctionType.Struct;
-            }
-
-            public void PopulateArguments(params Expression[] arguments)
-            {
-                PopulateArguments(arguments.ToList());
-            }
-
-            public void PopulateArguments(List<Expression> arguments)
-            {
-                if (Subtype != FunctionType.Struct)
-                    throw new InvalidOperationException("Cannot populate arguments of non-struct");
-
-                if (Arguments == null)
-                    Arguments = new List<Expression>();
-
-                Arguments.AddRange(arguments);
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                StringBuilder sb = new StringBuilder();
-                if (context.Statements.ContainsKey(FunctionBodyEntryBlock.Address.Value))
-                {
-                    FunctionDefinition def;
-                    var oldDecompilingStruct = context.DecompilingStruct;
-                    var oldReplacements = context.ArgumentReplacements;
-
-                    if (Subtype == FunctionType.Struct)
-                        context.DecompilingStruct = true;
-                    else
-                    {
-                        context.DecompilingStruct = false;
-                        sb.Append("function");
-                        if (IsStatement)
-                        {
-                            sb.Append(" ");
-
-                            // For further optimization, we could *probably* create a dictionary that's just flipped KVPs (assuming there are no dup. values).
-                            // Doing so would save the need for LINQ and what-not. Not that big of an issue, but still an option.
-                            Dictionary<string, UndertaleFunction> subFuncs = context.GlobalContext.Data.KnownSubFunctions;
-                            KeyValuePair<string, UndertaleFunction> kvp = subFuncs.FirstOrDefault(x => x.Value == Function);
-
-                            // If we found an associated sub-function, use the key as the name.
-                            if (kvp.Key != null)
-                                sb.Append(kvp.Key);
-                            else
-                            {
-                                //Attempt to find function names before going with the last functions' name
-                                bool gotFuncName = false;
-                                if (Function.Name.Content.StartsWith("gml_Script_"))
-                                {
-                                    string funcName = Function.Name.Content.Substring("gml_Script_".Length);
-                                    if (context.Statements[0].Any(x => x is AssignmentStatement && (x as AssignmentStatement).Destination.Var.Name.Content == funcName))
-                                    {
-                                        sb.Append(funcName);
-                                        gotFuncName = true;
-                                    }
-                                }
-                                if(!gotFuncName)
-                                    sb.Append((context.Statements[0].Last() as AssignmentStatement).Destination.Var.Name.Content);
-                            }
-                        }
-                        sb.Append("(");
-                        for (int i = 0; i < FunctionBodyCodeEntry.ArgumentsCount; ++i)
-                        {
-                            if (i != 0)
-                                sb.Append(", ");
-                            sb.Append("argument");
-                            sb.Append(i);
-                        }
-                        sb.Append(") ");
-                        if (Subtype == FunctionType.Constructor)
-                            sb.Append("constructor ");
-                        sb.Append("//");
-                        sb.Append(Function.Name.Content);
-                    }
-
-                    var statements = context.Statements[FunctionBodyEntryBlock.Address.Value];
-                    int numNotReturn = statements.FindAll(stmt => !(stmt is ReturnStatement)).Count;
-
-                    if (numNotReturn > 0 || Subtype != FunctionType.Struct)
-                    {
-                        sb.Append("\n");
-                        sb.Append(context.Indentation);
-                        sb.Append("{\n");
-                        context.IndentationLevel++;
-                        context.ArgumentReplacements = Arguments;
-
-                        int count = 0;
-                        foreach (Statement stmt in statements)
-                        {
-                            count++;
-                            if ((Subtype != FunctionType.Function && stmt is ReturnStatement) || (stmt is AssignmentStatement assign && assign.IsStructDefinition))
-                                continue;
-
-                            sb.Append(context.Indentation);
-
-                            // See #614
-                            // This is not the place to monkey patch this
-                            // issue, but it's like 2am and quite frankly
-                            // I don't care anymore.
-                            def = null;
-                            if (stmt is FunctionDefinition)
-                                def = stmt as FunctionDefinition;
-                            else if (stmt is TempVarAssignmentStatement reference && reference.Value is FunctionDefinition)
-                                def = reference.Value as FunctionDefinition;
-
-                            if (def?.Function == Function)
-                            {
-                                //sb.Append("// Error decompiling function: function contains its own declaration???\n");
-                                sb.Append("\n");
-                                break;
-                            }
-                            else
-                            {
-                                sb.Append(stmt.ToString(context));
-                                if (Subtype == FunctionType.Struct && count < numNotReturn)
-                                    sb.Append(",");
-                            }
-                            sb.Append("\n");
-                        }
-                        context.DecompilingStruct = oldDecompilingStruct;
-                        context.ArgumentReplacements = oldReplacements;
-                        context.IndentationLevel--;
-                        sb.Append(context.Indentation);
-                        sb.Append("}");
-                        if(!oldDecompilingStruct)
-                            sb.Append("\n");
-                    }
-                    else
-                        sb.Append("{}");
-                }
-                else
-                {
-                    sb.Append(Function.Name.Content);
-                }
-                return sb.ToString();
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return suggestedType;
-            }
-        }
-
-        // Represents a high-level function or script call.
-        public abstract class FunctionCall : Expression
-        {
-            internal UndertaleInstruction.DataType ReturnType;
-            internal List<Expression> Arguments;
-
-            protected FunctionCall(UndertaleInstruction.DataType returnType, List<Expression> args)
-            {
-                this.ReturnType = returnType;
-                this.Arguments = args;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                // Function calls are never duplication safe - they can have side effects
-                return false;
-            }
-        }
-
-        public class DirectFunctionCall : FunctionCall
-        {
-            internal string OverridenName = string.Empty;
-            internal UndertaleFunction Function;
-
-            public DirectFunctionCall(string overridenName, UndertaleFunction function, UndertaleInstruction.DataType returnType, List<Expression> args) : base(returnType, args)
-            {
-                this.OverridenName = overridenName;
-                this.Function = function;
-            }
-
-            public DirectFunctionCall(UndertaleFunction function, UndertaleInstruction.DataType returnType, List<Expression> args) : base(returnType, args)
-            {
-                this.Function = function;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                StringBuilder argumentString = new StringBuilder();
-
-                if (Function.Name.Content == "@@NewGMLObject@@") // Creating a new "object" via a constructor OR this is a struct definition
-                {
-                    context.currentFunction = this;
-
-                    string constructor;
-                    var actualArgs = Arguments.Skip(1).ToList();
-                    if (Arguments[0] is FunctionDefinition def)
-                    {
-                        if (def.Subtype == FunctionDefinition.FunctionType.Struct) // Struct moment
-                        {
-                            def.PopulateArguments(actualArgs);
-                            return def.ToString(context);
-                        }
-                        else
-                            constructor = def.FunctionBodyCodeEntry.Name.Content;
-                    }
-                    else
-                        constructor = Arguments[0].ToString(context);
-
-                    if (constructor.StartsWith("gml_Script_", StringComparison.InvariantCulture))
-                        constructor = constructor.Substring(11);
-                    if (constructor.EndsWith(context.TargetCode.Name.Content, StringComparison.InvariantCulture))
-                        constructor = constructor.Substring(0, constructor.Length - context.TargetCode.Name.Content.Length - 1);
-
-                    if (AssetTypeResolver.builtin_funcs.TryGetValue(constructor, out AssetIDType[] types))
-                    {
-                        int index = 0;
-                        foreach (var arg in actualArgs)
-                            arg.DoTypePropagation(context, types[index++]);
-                    }
-
-                    // Don't ask
-                    if (Arguments[0] is ExpressionCast cast &&
-                        cast.Argument is ExpressionConstant constant &&
-                        constant.Value is UndertaleInstruction.Reference<UndertaleFunction> reference)
-                    {
-                        var call = new DirectFunctionCall(reference.Target, ReturnType, actualArgs) {
-                            OverridenName = constructor
-                        };
-
-                        return "new " + call.ToString(context);
-                    }
-                    else
-                    {
-                        foreach (Expression exp in actualArgs)
-                        {
-                            context.currentFunction = this;
-                            if (argumentString.Length > 0)
-                                argumentString.Append(", ");
-                            argumentString.Append(exp.ToString(context));
-                        }
-                    }
-                    context.currentFunction = null;
-
-                    return String.Format("new {0}({1})", constructor, argumentString);
-                }
-                else
-                {
-                    foreach (Expression exp in Arguments)
-                    {
-                        context.currentFunction = this;
-                        if (argumentString.Length > 0)
-                            argumentString.Append(", ");
-                        argumentString.Append(exp.ToString(context));
-
-                    }
-                    context.currentFunction = null;
-
-                    if (Function.Name.Content == "@@NewGMLArray@@") // Inline array definitions
-                        return "[" + argumentString.ToString() + "]";
-
-                    return String.Format("{0}({1})", OverridenName != string.Empty ? OverridenName : Function.Name.Content, argumentString.ToString());
-                }
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                // Special case for these functions which don't have any purpose in decompiled code
-                if (Function?.Name?.Content == "@@This@@")
-                {
-                    return new ExpressionConstant(UndertaleInstruction.DataType.Variable, "self");
-                }
-                if (Function?.Name?.Content == "@@Other@@")
-                {
-                    return new ExpressionConstant(UndertaleInstruction.DataType.Variable, "other");
-                }
-                if (Function?.Name?.Content == "@@GetInstance@@")
-                {
-                    Statement res = Arguments[0];
-                    if (res is ExpressionCast cast)
-                        return cast.Argument;
-                    return res;
-                }
-                for (var i = 0; i < Arguments.Count; i++)
-                    Arguments[i] = Arguments[i]?.CleanExpression(context, block);
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                string funcName = OverridenName != string.Empty ? OverridenName : Function.Name.Content;
-                var script_code = context.GlobalContext.Data?.Scripts.ByName(funcName)?.Code;
-                if (script_code != null && !context.GlobalContext.ScriptArgsCache.ContainsKey(funcName))
-                {
-                    context.GlobalContext.ScriptArgsCache.Add(funcName, null); // stop the recursion from looping
-                    DecompileContext childContext;
-                    try
-                    {
-                        if (script_code.ParentEntry != null)
-                        {
-                            childContext = new DecompileContext(context.GlobalContext, script_code.ParentEntry);
-                            Dictionary<uint, Block> blocks = Decompiler.PrepareDecompileFlow(script_code.ParentEntry, new List<uint>() { script_code.Offset / 4 });
-                            Decompiler.DecompileFromBlock(childContext, blocks, blocks[script_code.Offset / 4]);
-                            Decompiler.DoTypePropagation(childContext, blocks); // TODO: This should probably put suggestedType through the "return" statement at the other end
-                        }
-                        else
-                        {
-                            childContext = new DecompileContext(context.GlobalContext, script_code);
-                            Dictionary<uint, Block> blocks = Decompiler.PrepareDecompileFlow(script_code, new List<uint>() { 0 });
-                            Decompiler.DecompileFromBlock(childContext, blocks, blocks[0]);
-                            Decompiler.DoTypePropagation(childContext, blocks); // TODO: This should probably put suggestedType through the "return" statement at the other end
-                        }
-                        context.GlobalContext.ScriptArgsCache[funcName] = new AssetIDType[15];
-                        for (int i = 0; i < 15; i++)
-                        {
-                            var v = childContext.assetTypes.Where((x) => x.Key.Name.Content == "argument" + i);
-                            context.GlobalContext.ScriptArgsCache[funcName][i] = v.Any() ? v.First().Value : AssetIDType.Other;
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        context.GlobalContext.DecompilerWarnings.Add("/*\nWARNING: Recursive script decompilation (for asset type resolution) failed for " + Function.Name.Content + "\n\n" + e.ToString() + "\n*/");
-                    }
-                }
-
-                AssetIDType[] args = new AssetIDType[Arguments.Count];
-                AssetTypeResolver.AnnotateTypesForFunctionCall(funcName, args, context, this);
-                for (var i = 0; i < Arguments.Count; i++)
-                    Arguments[i].DoTypePropagation(context, args[i]);
-
-                return suggestedType; // TODO: maybe we should handle returned values too?
-            }
-        }
-
-        public class IndirectFunctionCall : FunctionCall
-        {
-            internal Expression FunctionThis;
-            internal Expression Function;
-
-            public IndirectFunctionCall(Expression func_this, Expression func, UndertaleInstruction.DataType returnType, List<Expression> args) : base(returnType, args)
-            {
-                this.FunctionThis = func_this;
-                this.Function = func;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                StringBuilder argumentString = new StringBuilder();
-                foreach (Expression exp in Arguments)
-                {
-                    if (argumentString.Length > 0)
-                        argumentString.Append(", ");
-                    argumentString.Append(exp.ToString(context));
-                }
-
-                if (Function is FunctionDefinition)
-                    return String.Format("{0}({1})", Function.ToString(context), argumentString.ToString());
-
-                return String.Format("{0}.{1}({2})", FunctionThis.ToString(context), Function.ToString(context), argumentString.ToString());
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                FunctionThis = (Expression)FunctionThis?.CleanStatement(context, block);
-                Function = (Expression)Function?.CleanStatement(context, block);
-                for (var i = 0; i < Arguments.Count; i++)
-                    Arguments[i] = Arguments[i]?.CleanExpression(context, block);
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                FunctionThis.DoTypePropagation(context, AssetIDType.GameObject);
-                Function.DoTypePropagation(context, suggestedType);
-                AssetIDType[] args = new AssetIDType[Arguments.Count];
-                for (var i = 0; i < Arguments.Count; i++)
-                    Arguments[i].DoTypePropagation(context, args[i]);
-
-                return suggestedType;
-            }
-        }
-
-        // Represents a variable in an expression, of any type.
-        public class ExpressionVar : Expression
-        {
-            public UndertaleVariable Var;
-            public Expression InstType; // UndertaleInstruction.InstanceType
-            public UndertaleInstruction.VariableType VarType;
-            public List<Expression> ArrayIndices = null;
-            public UndertaleInstruction.Opcode Opcode;
-
-            public ExpressionVar(UndertaleVariable var, Expression instType, UndertaleInstruction.VariableType varType)
-            {
-                Var = var;
-                InstType = instType;
-                VarType = varType;
-            }
-
-            internal override bool IsDuplicationSafe()
-            {
-                bool res = (InstType?.IsDuplicationSafe() ?? true);
-
-                if (ArrayIndices == null)
-                    return res;
-                foreach (Expression e in ArrayIndices)
-                    res &= (e?.IsDuplicationSafe() ?? true);
-
-                return res;
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                if (Var.Name?.Content == "$$$$temp$$$$" && context.CompilerTempVar != null)
-                {
-                    block.Statements.Remove(context.CompilerTempVar);
-                    return context.CompilerTempVar.Value.CleanStatement(context, block);
-                }
-
-                InstType = InstType?.CleanExpression(context, block);
-                if (ArrayIndices == null)
-                    return this;
-                foreach (Expression e in ArrayIndices)
-                    e?.CleanExpression(context, block);
-                return this;
-            }
-
-            public static Tuple<Expression, Expression> Decompile2DArrayIndex(Expression index)
-            {
-                Expression ind1 = index;
-                Expression ind2 = null;
-                if (ind1 is ExpressionTwo && (ind1 as ExpressionTwo).Opcode == UndertaleInstruction.Opcode.Add)
-                {
-                    var arg1 = (ind1 as ExpressionTwo).Argument1;
-                    var arg2 = (ind1 as ExpressionTwo).Argument2;
-                    if (arg1 is ExpressionTwo && (arg1 as ExpressionTwo).Opcode == UndertaleInstruction.Opcode.Mul)
-                    {
-                        var arg11 = (arg1 as ExpressionTwo).Argument1;
-                        var arg12 = (arg1 as ExpressionTwo).Argument2;
-                        if (arg12 is ExpressionConstant && (arg12 as ExpressionConstant).Value is int && (int)(arg12 as ExpressionConstant).Value == 32000)
-                        {
-                            ind1 = arg11;
-                            ind2 = arg2;
-                        }
-                    }
-                }
-                return new Tuple<Expression, Expression>(ind1, ind2);
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                string name = Var.Name.Content;
-                if (ArrayIndices != null)
-                {
-                    if (DecompileContext.GMS2_3 == true)
-                    {
-                        if (name == "argument" && context.DecompilingStruct && context.ArgumentReplacements != null && ArrayIndices.Count == 1)
-                        {
-                            var replacements = context.ArgumentReplacements;
-                            if (int.TryParse(ArrayIndices[0].ToString(context), out int index) && index >= 0 && index < replacements.Count && this != replacements[index])
-                                return replacements[index].ToString(context);
-                        }
-                        foreach (Expression e in ArrayIndices)
-                            name += "[" + e.ToString(context) + "]";
-
-                    }
-                    else
-                    {
-                        if (ArrayIndices.Count == 2 && ArrayIndices[0] != null && ArrayIndices[1] != null)
-                            name += "[" + ArrayIndices[0].ToString(context) + ", " + ArrayIndices[1].ToString(context) + "]";
-                        else if (ArrayIndices[0] != null)
-                            name += "[" + ArrayIndices[0].ToString(context) + "]";
-                    }
-                }
-
-                // NOTE: The "var" prefix is handled in Decompiler.Decompile.
-
-                if (VarType == UndertaleInstruction.VariableType.Instance)
-                {
-                    if (InstType is ExpressionConstant c)
-                    {
-                        int? val = ExpressionConstant.ConvertToInt(c.Value);
-                        if (val == null)
-                            throw new InvalidOperationException("Unable to parse the instance ID to int");
-                        // TODO: This is a reference to an object instance in the room. Resolving these is non-trivial since you don't exactly have a reference to the room where this script is used when decompiling...
-                        return (val + 100000) + "." + name;
-                    }
-                    else throw new InvalidOperationException("Instance variable type used with non-const InstType"); // TODO: can this happen?
-                }
-                if (InstType is ExpressionConstant constant) // Only use "global." and "other.", not "self." or "local.". GMS doesn't recognize those.
-                {
-                    string prefix = InstType.ToString(context) + ".";
-                    if (!(constant.Value is Int64))
-                    {
-                        int? val = ExpressionConstant.ConvertToInt(constant.Value);
-                        if (val != null)
-                        {
-                            if (constant.AssetType == AssetIDType.GameObject && val < 0)
-                            {
-                                UndertaleInstruction.InstanceType instanceType = (UndertaleInstruction.InstanceType)val;
-                                prefix = (instanceType == UndertaleInstruction.InstanceType.Global || instanceType == UndertaleInstruction.InstanceType.Other) ? prefix.ToLower(CultureInfo.InvariantCulture) : "";
-                            }
-                        }
-                    }
-                    return prefix + name;
-                }
-                else if (InstType is ExpressionCast cast && !(cast.Argument is ExpressionVar))
-                {
-                    return "(" + InstType.ToString(context) + ")." + name; // Make sure to put parentheses around these cases
-                }
-
-                return InstType.ToString(context) + "." + name;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                InstType?.DoTypePropagation(context, AssetIDType.GameObject);
-                if (ArrayIndices != null)
-                {
-                    foreach (Expression e in ArrayIndices)
-                        e?.DoTypePropagation(context, AssetIDType.Other);
-                }
-
-                AssetIDType current = context.assetTypes.ContainsKey(Var) ? context.assetTypes[Var] : AssetIDType.Other;
-                if (current == AssetIDType.Other && suggestedType != AssetIDType.Other)
-                    current = suggestedType;
-                AssetIDType builtinSuggest = AssetTypeResolver.AnnotateTypeForVariable(context, Var.Name.Content);
-                if (builtinSuggest != AssetIDType.Other)
-                    current = builtinSuggest;
-
-                if ((VarType != UndertaleInstruction.VariableType.Array || (ArrayIndices != null && !(ArrayIndices[0] is ExpressionConstant))))
-                    context.assetTypes[Var] = current; // This is a messy fix to arrays messing up exported variable types.
-                return current;
-            }
-        }
-
-        // Represents a with statement beginning (pushing to the env stack).
-        // This is not seen in high-level output.
-        public class PushEnvStatement : Statement
-        {
-            public Expression NewEnv;
-
-            public PushEnvStatement(Expression newEnv)
-            {
-                this.NewEnv = newEnv;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                return "pushenv " + NewEnv;
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                NewEnv = NewEnv?.CleanExpression(context, block);
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                NewEnv.DoTypePropagation(context, AssetIDType.GameObject);
-                return suggestedType;
-            }
-        }
-
-        // Represents a with statement ending (popping from or clearing the env stack).
-        // This is not seen in high-level output.
-        public class PopEnvStatement : Statement
-        {
-            public override string ToString(DecompileContext context)
-            {
-                return "popenv";
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                return this;
-            }
-
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                return suggestedType;
-            }
-        }
-
+        
         static int GetTypeSize(UndertaleInstruction.DataType type)
         {
             switch (type)
@@ -2923,505 +910,6 @@ namespace UndertaleModLib.Decompiler
                     block.nextBlockFalse.entryPoints.Add(block);
             }
             return blockByAddress;
-        }
-
-        public abstract class HLStatement : Statement
-        {
-            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
-            {
-                throw new NotImplementedException();
-            }
-        };
-
-        public class BlockHLStatement : HLStatement
-        {
-            public List<Statement> Statements = new List<Statement>();
-
-            public string ToString(DecompileContext context, bool canSkipBrackets = true, bool forceSkipBrackets = false)
-            {
-                context.IndentationLevel++;
-                if (canSkipBrackets && CanSkipBrackets(this))
-                {
-                    string res = DecompileContext.Indent + Statements[0].ToString(context);
-                    context.IndentationLevel--;
-                    return res;
-                }
-                else
-                {
-                    StringBuilder sb = new StringBuilder();
-                    if (!forceSkipBrackets)
-                        sb.Append("{\n");
-                    foreach (var stmt in Statements)
-                    {
-                        if (stmt is AssignmentStatement assign && assign.IsStructDefinition)
-                            continue;
-
-                        sb.Append(context.Indentation);
-                        string resultStr = stmt.ToString(context);
-                        sb.Append(resultStr).Append('\n');
-                    }
-                    context.IndentationLevel--;
-                    if (!forceSkipBrackets)
-                        sb.Append(context.Indentation + "}");
-                    return sb.ToString().Trim('\n');
-                }
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                for (var i = 0; i < Statements.Count; i++)
-                {
-                    var count = Statements.Count;
-                    var Result = Statements[i]?.CleanStatement(context, this); // Yes, this uses "this" and not "block".
-                    i -= (count - Statements.Count); // If removed.
-                    Statements[i] = Result;
-                }
-                return this;
-            }
-
-            public BlockHLStatement CleanBlockStatement(DecompileContext context)
-            {
-                return CleanStatement(context, null) as BlockHLStatement;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                return ToString(context, true);
-            }
-
-            private static bool CanSkipBrackets(BlockHLStatement blockStatement)
-            {
-                if (blockStatement == null || blockStatement.Statements.Count != 1)
-                    return false; // Nope! Need brackets!
-
-                Statement statement = blockStatement.Statements[0];
-                return !(statement is IfHLStatement || statement is LoopHLStatement || statement is HLSwitchStatement || statement is WithHLStatement); // Nesting these can cause issues.
-            }
-        };
-
-        public class IfHLStatement : HLStatement
-        {
-            public Expression condition;
-            public BlockHLStatement trueBlock;
-            public List<ValueTuple<Expression, BlockHLStatement>> elseConditions = new List<ValueTuple<Expression, BlockHLStatement>>();
-            public BlockHLStatement falseBlock;
-
-            public bool HasElseIf { get => elseConditions != null && elseConditions.Count > 0; }
-            public bool HasElse { get => falseBlock != null && falseBlock.Statements.Count > 0; }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                condition = condition?.CleanExpression(context, block);
-                trueBlock = trueBlock?.CleanBlockStatement(context);
-                falseBlock = falseBlock?.CleanBlockStatement(context);
-
-                int myIndex = block.Statements.IndexOf(this);
-
-                // Prevents if (tempvar - 1), when it should be if (tempvar)
-                if ((condition as ExpressionCast)?.Argument is ExpressionTwo && myIndex > 0)
-                {
-                    ExpressionTwo conditionExpression = ((ExpressionCast)condition).Argument as ExpressionTwo;
-                    Statement lastStatement = block.Statements[myIndex - 1];
-
-                    if (conditionExpression.Argument1 is ExpressionTempVar tempVar && lastStatement is TempVarAssignmentStatement statement && conditionExpression.Argument2 is ExpressionConstant
-                        && tempVar.Var.Var == statement.Var.Var)
-                        condition = conditionExpression.Argument1;
-                }
-
-                // Use if -> else if, instead of nesting ifs.
-                while (falseBlock.Statements.Count == 1 && falseBlock.Statements[0] is IfHLStatement nestedIf) // The condition of one if statement.
-                {
-                    elseConditions.Add(new ValueTuple<Expression, BlockHLStatement>(nestedIf.condition, nestedIf.trueBlock));
-                    elseConditions.AddRange(nestedIf.elseConditions);
-                    falseBlock = nestedIf.falseBlock;
-                }
-
-                // Collapse conditions into && + || + ternary.
-                if (HasElse && !HasElseIf && trueBlock.Statements.Count == 1 && falseBlock.Statements.Count == 1)
-                {
-                    TempVarAssignmentStatement trueAssign = trueBlock.Statements[0] as TempVarAssignmentStatement;
-                    TempVarAssignmentStatement falseAssign = falseBlock.Statements[0] as TempVarAssignmentStatement;
-
-                    if (trueAssign != null && falseAssign != null && trueAssign.Var.Var == falseAssign.Var.Var)
-                    {
-                        TempVarAssignmentStatement newAssign;
-                        if (TestNumber(trueAssign.Value, 1) && (falseAssign.Var.Var.Type == UndertaleInstruction.DataType.Boolean || falseAssign.Value.Type == UndertaleInstruction.DataType.Boolean))
-                            newAssign = new TempVarAssignmentStatement(trueAssign.Var, new ExpressionTwoSymbol("||", UndertaleInstruction.DataType.Boolean, condition, falseAssign.Value));
-                        else if (TestNumber(falseAssign.Value, 0) && (trueAssign.Var.Var.Type == UndertaleInstruction.DataType.Boolean || trueAssign.Value.Type == UndertaleInstruction.DataType.Boolean))
-                            newAssign = new TempVarAssignmentStatement(trueAssign.Var, new ExpressionTwoSymbol("&&", UndertaleInstruction.DataType.Boolean, condition, trueAssign.Value));
-                        else
-                            newAssign = new TempVarAssignmentStatement(trueAssign.Var, new ExpressionTernary(trueAssign.Value.Type, condition, trueAssign.Value, falseAssign.Value));
-
-                        context.TempVarMap[newAssign.Var.Var.Name] = newAssign;
-                        return newAssign;
-                    }
-                }
-
-                // Create repeat loops.
-                if (HasElse && !HasElseIf && trueBlock.Statements.Count == 0 && falseBlock.Statements.Count == 1 && falseBlock.Statements[0] is LoopHLStatement
-                        && condition is ExpressionCompare && myIndex > 0 && block.Statements[myIndex - 1] is TempVarAssignmentStatement)
-                {
-                    ExpressionCompare compareCondition = condition as ExpressionCompare;
-                    LoopHLStatement loop = falseBlock.Statements[0] as LoopHLStatement;
-                    TempVarAssignmentStatement priorAssignment = block.Statements[myIndex - 1] as TempVarAssignmentStatement;
-                    Expression startValue = priorAssignment.Value;
-
-                    List<Statement> loopCode = loop.Block.Statements;
-                    if (priorAssignment != null &&
-                        loop.IsWhileLoop &&
-                        loop.Condition == null &&
-                        //loopCode.Count > 2 &&
-                        compareCondition.Opcode == UndertaleInstruction.ComparisonType.LTE &&
-                        TestNumber(compareCondition.Argument2, 0) &&
-                        compareCondition.Argument1.ToString(context) == startValue.ToString(context))
-                    {
-                        TempVarAssignmentStatement repeatAssignment = null;
-                        IfHLStatement loopCheckStatement = null;
-                        bool hasBreak = false;
-                        List<Statement> insideElseBlock = null;
-
-                        if (loopCode.Count > 2)
-                        {
-                            repeatAssignment = loopCode[loopCode.Count - 2] as TempVarAssignmentStatement;
-                            loopCheckStatement = loopCode[loopCode.Count - 1] as IfHLStatement;
-                        }
-
-                        if ((repeatAssignment == null || loopCheckStatement == null) &&
-                            loopCode[loopCode.Count - 1] is IfHLStatement wrapperIfStatement &&
-                            wrapperIfStatement.HasElse
-                           ) // single-level break detection
-                        {
-                            insideElseBlock = wrapperIfStatement.falseBlock.Statements;
-                            wrapperIfStatement.trueBlock.Statements.Add(new BreakHLStatement());
-                            if (insideElseBlock.Count > 2)
-                            {
-                                repeatAssignment = insideElseBlock[insideElseBlock.Count - 2] as TempVarAssignmentStatement;
-                                loopCheckStatement = insideElseBlock[insideElseBlock.Count - 1] as IfHLStatement;
-                                hasBreak = true;
-                            }
-                        }
-
-                        if (repeatAssignment != null && loopCheckStatement != null)
-                        { // tempVar = (tempVar -1); -> if (tempVar) continue -> break
-
-                            // if (tempVar) {continue} else {empty}
-
-                            if (loopCheckStatement.trueBlock.Statements.Count == 1
-                                && !loopCheckStatement.HasElse
-                                && !loopCheckStatement.HasElseIf
-                                && loopCheckStatement.trueBlock.Statements[0] is ContinueHLStatement
-                                && loopCheckStatement.condition.ToString(context) == repeatAssignment.Value.ToString(context))
-                            {
-                                (hasBreak ? insideElseBlock : loopCode).Remove(repeatAssignment);
-                                (hasBreak ? insideElseBlock : loopCode).Remove(loopCheckStatement);
-                                block.Statements.Remove(priorAssignment);
-
-                                loop.RepeatStartValue = startValue;
-                                return loop;
-                            }
-                        }
-                    }
-                }
-
-                for (int i = 0; i < elseConditions.Count; i++)
-                {
-                    var pair = elseConditions[i];
-                    pair.Item1 = pair.Item1?.CleanExpression(context, block);
-                    pair.Item2 = pair.Item2?.CleanBlockStatement(context);
-                }
-
-
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                StringBuilder sb = new StringBuilder();
-                string cond;
-                if (condition is ExpressionCompare)
-                    cond = (condition as ExpressionCompare).ToStringWithParen(context);
-                else
-                    cond = condition.ToString(context);
-                sb.Append("if " + cond + "\n");
-                sb.Append(context.Indentation + trueBlock.ToString(context));
-
-                foreach (ValueTuple<Expression, BlockHLStatement> tuple in elseConditions)
-                {
-                    if (tuple.Item1 is ExpressionCompare)
-                        cond = (tuple.Item1 as ExpressionCompare).ToStringWithParen(context);
-                    else
-                        cond = tuple.Item1.ToString(context);
-                    sb.Append("\n" + context.Indentation + "else if " + cond + "\n");
-                    sb.Append(context.Indentation + tuple.Item2.ToString(context));
-                }
-
-                if (HasElse)
-                {
-                    sb.Append("\n" + context.Indentation + "else\n");
-                    sb.Append(context.Indentation + falseBlock.ToString(context));
-                }
-                return sb.ToString();
-            }
-        };
-
-        public class LoopHLStatement : HLStatement
-        {
-            // Loop block
-            public BlockHLStatement Block;
-
-            // While / for condition
-            public Statement Condition;
-
-            // While loop
-            public bool IsWhileLoop { get => !IsForLoop && !IsRepeatLoop && !IsDoUntilLoop; }
-            public bool IsDoUntilLoop;
-
-            // For loop
-            public bool IsForLoop { get => InitializeStatement != null && StepStatement != null && Condition != null; }
-            public AssignmentStatement InitializeStatement;
-            public AssignmentStatement StepStatement;
-
-            // Repeat loop
-            public bool IsRepeatLoop { get => RepeatStartValue != null; }
-            public Statement RepeatStartValue;
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                int myIndex = block.Statements.IndexOf(this);
-                Block = Block?.CleanBlockStatement(context);
-                Condition = Condition?.CleanStatement(context, block);
-                InitializeStatement = InitializeStatement?.CleanStatement(context, block) as AssignmentStatement;
-                StepStatement = StepStatement?.CleanStatement(context, block) as AssignmentStatement;
-                RepeatStartValue = RepeatStartValue?.CleanStatement(context, block);
-
-                if (IsWhileLoop)
-                {
-                    // While loops have conditions.
-                    if (Block.Statements.Count == 1)
-                    {
-                        Statement firstStatement = Block.Statements[0];
-
-                        if (firstStatement is IfHLStatement)
-                        {
-                            IfHLStatement ifStatement = (IfHLStatement)firstStatement;
-                            if (ifStatement.falseBlock is BlockHLStatement statement && statement.Statements.Count == 0 && !ifStatement.HasElseIf)
-                            {
-                                Condition = ifStatement.condition;
-                                Block.Statements.Remove(firstStatement); // Remove if statement.
-                                Block.Statements.InsertRange(0, ifStatement.trueBlock.Statements); // Add if contents.
-                            }
-                        }
-                    }
-
-                    // If it's been shoved into an if else, take it out of there.
-                    //TODO: This is disabled until further notice. The problem is that for loops that do this can use continues. The issue there is we don't have an easy way to remove the increment from the decompile at the moment.
-                    /*if (Block.Statements.Count > 0 && Block.Statements.Last() is IfHLStatement)
-                    {
-                        IfHLStatement ifStatement = Block.Statements.Last() as IfHLStatement;
-                        BlockHLStatement blockStatement = ifStatement.falseBlock as BlockHLStatement;
-
-                        if (blockStatement != null && blockStatement.Statements.Count > 0 && blockStatement.Statements.Last() is ContinueHLStatement)
-                        {
-                            blockStatement.Statements.Remove(blockStatement.Statements.Last());
-                            Block.Statements.AddRange(blockStatement.Statements);
-                            ifStatement.falseBlock.Statements.Clear();
-                        }
-                    }*/
-
-                    // Remove redundant continues at the end of the loop.
-                    if (Block.Statements.Count > 0)
-                    {
-                        Statement lastStatement = Block.Statements.Last();
-                        if (lastStatement is ContinueHLStatement)
-                            Block.Statements.Remove(lastStatement);
-                    }
-
-                    // Convert into a for loop.
-                    if (myIndex > 0 && block.Statements[myIndex - 1] is AssignmentStatement assignment
-                        && Block.Statements.Count > 0 && Block.Statements.Last() is AssignmentStatement increment
-                        && Condition is ExpressionCompare compare)
-                    {
-                        UndertaleVariable variable = assignment.Destination.Var;
-
-                        if (((compare.Argument1 is ExpressionVar exprVar && (exprVar.Var == variable)) || (compare.Argument2 is ExpressionVar exprVar2 && (exprVar2.Var == variable))) && increment.Destination.Var == variable)
-                        {
-                            block.Statements.Remove(assignment);
-                            InitializeStatement = assignment;
-                            Block.Statements.Remove(increment);
-                            StepStatement = increment;
-                        }
-                    }
-
-                    if (Condition == null)
-                    {
-                        if (Block.Statements.Last() is IfHLStatement)
-                        {
-                            IfHLStatement ifStatement = Block.Statements.Last() as IfHLStatement;
-                            if (ifStatement.trueBlock.Statements.Count == 0 && ifStatement.falseBlock.Statements.Count == 1 && ifStatement.falseBlock.Statements[0] is ContinueHLStatement && ifStatement.elseConditions.Count == 0)
-                            {
-                                IsDoUntilLoop = true;
-                                Condition = ifStatement.condition;
-                                Block.Statements.Remove(ifStatement);
-                            }
-                        }
-                    }
-                }
-
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                if (IsRepeatLoop)
-                {
-                    bool needsParen = RepeatStartValue is ExpressionConstant || RepeatStartValue is ExpressionCompare;
-                    return "repeat " + (needsParen ? "(" : "") + RepeatStartValue.ToString(context) + (needsParen ? ")" : "") + "\n" + context.Indentation + Block.ToString(context);
-                }
-
-                if (IsForLoop)
-                {
-                    string conditionStr = Condition.ToString(context); // Cut off parenthesis for the condition.
-                    if (conditionStr.StartsWith("(", StringComparison.InvariantCulture) && conditionStr.EndsWith(")", StringComparison.InvariantCulture))
-                        conditionStr = conditionStr.Substring(1, conditionStr.Length - 2);
-
-                    return "for (" + InitializeStatement.ToString(context) + "; " + conditionStr + "; " + StepStatement.ToString(context) + ")\n" + context.Indentation + Block.ToString(context);
-                }
-
-                string cond;
-                if (Condition is ExpressionCompare)
-                    cond = (Condition as ExpressionCompare).ToStringWithParen(context);
-                else if (IsDoUntilLoop)
-                    cond = Condition?.ToString(context) ?? "(false)";
-                else
-                    cond = Condition != null ? Condition.ToString(context) : "(true)";
-
-                if (IsDoUntilLoop)
-                    return "do\n" + context.Indentation + Block.ToString(context, false) + " until " + cond + ";";
-
-                return "while " + cond + "\n" + context.Indentation + Block.ToString(context);
-            }
-        };
-
-        public class ContinueHLStatement : HLStatement
-        {
-            public override string ToString(DecompileContext context)
-            {
-                return "continue";
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                return this;
-            }
-        }
-
-        public class BreakHLStatement : HLStatement
-        {
-            public override string ToString(DecompileContext context)
-            {
-                return "break";
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                return this;
-            }
-        }
-
-        public class WithHLStatement : HLStatement
-        {
-            public Expression NewEnv;
-            public BlockHLStatement Block;
-
-            public override string ToString(DecompileContext context)
-            {
-                return "with (" + NewEnv.ToString(context) + ")\n" + context.Indentation + Block.ToString(context);
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                NewEnv = NewEnv?.CleanExpression(context, block);
-                Block = Block?.CleanBlockStatement(context);
-                return this;
-            }
-        }
-
-        public class HLSwitchStatement : HLStatement
-        {
-            private Expression SwitchExpression;
-            private List<HLSwitchCaseStatement> Cases;
-
-            public HLSwitchStatement(Expression switchExpression, List<HLSwitchCaseStatement> cases)
-            {
-                this.SwitchExpression = switchExpression;
-                this.Cases = cases;
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                SwitchExpression = SwitchExpression?.CleanExpression(context, block);
-                for (var i = 0; i < Cases.Count; i++)
-                    Cases[i] = Cases[i]?.CleanStatement(context, block) as HLSwitchCaseStatement;
-
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                StringBuilder sb = new StringBuilder();
-                sb.Append("switch " + SwitchExpression.ToString(context) + "\n");
-                sb.Append(context.Indentation + "{\n");
-                context.IndentationLevel++;
-                foreach (var casee in Cases)
-                {
-                    sb.Append(context.Indentation + casee.ToString(context));
-                    sb.Append('\n');
-                }
-                context.IndentationLevel--;
-                sb.Append(context.Indentation + "}\n");
-                return sb.ToString();
-            }
-        }
-
-        public class HLSwitchCaseStatement : HLStatement
-        {
-            public List<Expression> CaseExpressions;
-            public BlockHLStatement Block;
-
-            public HLSwitchCaseStatement(List<Expression> caseExpressions, BlockHLStatement block)
-            {
-                DebugUtil.Assert(caseExpressions.Count > 0, "Switch statement lacks any cases.");
-                this.CaseExpressions = caseExpressions;
-                this.Block = block;
-            }
-
-            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
-            {
-                for (var i = 0; i < CaseExpressions.Count; i++)
-                    CaseExpressions[i] = CaseExpressions[i]?.CleanExpression(context, block);
-                Block = Block?.CleanBlockStatement(context);
-                return this;
-            }
-
-            public override string ToString(DecompileContext context)
-            {
-                StringBuilder sb = new StringBuilder();
-                for (int i = 0; i < CaseExpressions.Count; i++)
-                {
-                    Expression caseExpr = CaseExpressions[i];
-                    if (i != 0)
-                        sb.Append(context.Indentation);
-                    if (caseExpr != null)
-                        sb.Append("case " + caseExpr.ToString(context) + ":\n");
-                    else
-                        sb.Append("default:\n");
-                }
-                if (Block.Statements.Count > 0)
-                {
-                    sb.Append(Block.ToString(context, false, true));
-                }
-                return sb.ToString();
-            }
         }
 
         // Based on http://www.backerstreet.com/decompiler/loop_analysis.php
