+++ UndertaleModLib/Decompiler/Instructions/Decompiler.ContinueHLStatement.cs
@@ -1,0 +1,17 @@
+namespace UndertaleModLib.Decompiler;
+
+public static partial class Decompiler
+{
+    public class ContinueHLStatement : HLStatement
+    {
+        public override string ToString(DecompileContext context)
+        {
+            return "continue";
+        }
+
+        public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
+        {
+            return this;
+        }
+    }
+}
